requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
